Enhance Azure credential configuration
Enhanced the configuration of DefaultAzureCredentialOptions in ServiceCollectionExtensions.cs within the CallRecordInsights.Extensions namespace. Added a conditional check to set the AuthorityHost property to a custom URI if identityOptions.Authority is not null. Also, updated the AdditionallyAllowedTenants collection by adding a wildcard character, "*", to allow credentials to be used across multiple tenants for increased flexibility.